Move parts of action_text and the Text action into a new file
